"Finalized the demo of Always"
In the AlwaysBindingExample.java, a constantly updating clock is given
as an example. It's updated twice every second, as dictated in the
creation of a timer in the Always.java class. The interval is within
the Always.java class, but that should be fixed when I figure out how
to handle multiple parameters given at the binding line.

Always.java and AlwaysBindingExample.java are the only two files needed
to run this, in addition to changes made to the appropiate files
inbetween, as it is not currently properly extended as a binding but
rather placed together with all of the proper bindings. This will
change and be properly settled in an upcoming commit.

Running the example program, given Always.java is in the right area and
has been manually added to the other files, should result in a JFrame
clock that updates without user intervention.

Since the master branch relies on an outdated verson of guava, I've
included my entire eclipse workspace to ease in patching library
dependencies.

The issue with binding to only specific types has been resolved.