Save taking in the delay parameter at the bind line, it's ~finalized.
Within Clock/src/alwaysBindingDemo is a working demo of the Always
binding, which relies on always calling a controller method. This is
useful in having the view dynamically update without user intervention,
such as to display the current time as in the given example.

The major change has been the officializing of this binding extension,
which boils down the the following three classes within the
aforementioned folder
1) Always.java - the actual binding
2) AlwaysExtensionBindings.java - An extension of the Bindings class
that serves to allow the Always extension without having to change the
source code of any files. Be sure that any implementing classes use
"AlwaysExtensionBindings.alwaysExtensionBindings()" as the source for
bindings, not the default.
3) AlwaysBindingExample.java - The program which shows how this binding
could prove useful.